+++ changelog.md
@@ -1,8 +1,9 @@
-* Added Iron, Gold, Redstone, Lapis, Emerald, Diamond & Netherite Bricks
+* Added **Iron, Gold, Redstone, Lapis, Emerald, Diamond & Netherite Bricks**
     * Can be converted back into ingots/gems (slabs give 4)
     * Have the same properties as the ore block assigned (e.g. Redstone Bricks provides redstone power)
     * Walls cannot be used as beacon base.
-
+* Added **Prismarine Tiles** - made by 4 using 2 prismarine shards & 2 dark prismarine
+* Added **Obsidian Pressure Plate** that can only be triggered by a player.
 
 * Tweaked most glazed terracotta pillars textures.
     * Magenta, Purple & Cyan have a completly different style.